refactor: split hardware sheet screens into files
