+++ app/src/main/java/to/bitkit/ui/sheets/HardwareSheet.kt
@@ -1,53 +1,17 @@
 package to.bitkit.ui.sheets
 
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.navigationBarsPadding
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import kotlinx.serialization.Serializable
-import to.bitkit.R
-import to.bitkit.ui.components.BodyM
-import to.bitkit.ui.components.BottomSheetPreview
-import to.bitkit.ui.components.Display
-import to.bitkit.ui.components.FillHeight
-import to.bitkit.ui.components.KEY_DELETE
-import to.bitkit.ui.components.NumberPad
-import to.bitkit.ui.components.PrimaryButton
-import to.bitkit.ui.components.SecondaryButton
 import to.bitkit.ui.components.Sheet
 import to.bitkit.ui.components.SheetSize
-import to.bitkit.ui.components.VerticalSpacer
-import to.bitkit.ui.scaffold.SheetTopBar
 import to.bitkit.ui.shared.modifiers.sheetHeight
-import to.bitkit.ui.shared.util.gradientBackground
-import to.bitkit.ui.theme.AppThemeSurface
-import to.bitkit.ui.theme.Colors
 import to.bitkit.ui.utils.composableWithDefaultTransitions
-import to.bitkit.ui.utils.withAccent
 
 /**
  * Entry point for the hardware-wallet connect flow opened from the home suggestion
@@ -68,7 +32,6 @@ fun HardwareSheet(
         modifier = Modifier
             .fillMaxWidth()
             .sheetHeight(SheetSize.LARGE)
-            .gradientBackground()
             .testTag("hardware_sheet")
     ) {
         NavHost(
@@ -76,134 +39,15 @@ fun HardwareSheet(
             startDestination = sheet.route,
         ) {
             composableWithDefaultTransitions<HardwareRoute.Intro> {
-                HardwareIntro(onClose = onDismiss)
+                HwIntroSheet(onDismiss)
             }
             composableWithDefaultTransitions<HardwareRoute.PairingCode> {
-                HardwarePairing(
+                HwPairSheet(
                     onSubmit = onSubmitPairingCode,
                     onCancel = onCancelPairingCode,
                 )
             }
         }
-    }
-}
-
-@Composable
-private fun HardwareIntro(onClose: () -> Unit) {
-    Column(modifier = Modifier.fillMaxSize()) {
-        SheetTopBar(titleText = stringResource(R.string.hardware__intro_title))
-        BoxWithConstraints(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
-        ) {
-            val imageSize = maxWidth * INTRO_IMAGE_SIZE_RATIO
-            val staggerY = maxWidth * INTRO_IMAGE_STAGGER_RATIO
-            Image(
-                painter = painterResource(R.drawable.trezor),
-                contentDescription = null,
-                modifier = Modifier
-                    .size(imageSize)
-                    .align(Alignment.CenterStart)
-                    .offset(x = -maxWidth * INTRO_TREZOR_BLEED_RATIO, y = staggerY)
-            )
-            Image(
-                painter = painterResource(R.drawable.ledger),
-                contentDescription = null,
-                modifier = Modifier
-                    .size(imageSize)
-                    .align(Alignment.CenterEnd)
-                    .offset(x = maxWidth * INTRO_LEDGER_BLEED_RATIO, y = -staggerY)
-            )
-        }
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 32.dp)
-                .navigationBarsPadding()
-        ) {
-            Display(stringResource(R.string.hardware__intro_header).withAccent(accentColor = Colors.Blue))
-            VerticalSpacer(8.dp)
-            BodyM(stringResource(R.string.hardware__intro_text), color = Colors.White64)
-            VerticalSpacer(32.dp)
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(16.dp),
-                modifier = Modifier.fillMaxWidth(),
-            ) {
-                SecondaryButton(
-                    text = stringResource(R.string.common__cancel),
-                    onClick = onClose,
-                    modifier = Modifier.weight(1f)
-                )
-                PrimaryButton(
-                    text = stringResource(R.string.common__continue),
-                    onClick = {},
-                    enabled = false,
-                    modifier = Modifier.weight(1f)
-                )
-            }
-            VerticalSpacer(16.dp)
-        }
-    }
-}
-
-@Composable
-private fun HardwarePairing(
-    onSubmit: (String) -> Unit,
-    onCancel: () -> Unit,
-) {
-    var code by remember { mutableStateOf("") }
-    var submitted by remember { mutableStateOf(false) }
-
-    // Dismissing the sheet without submitting (e.g. swipe down) cancels the pending
-    // pairing request so the connect attempt does not hang until its timeout.
-    DisposableEffect(Unit) {
-        onDispose { if (!submitted) onCancel() }
-    }
-
-    Column(modifier = Modifier.fillMaxSize()) {
-        SheetTopBar(titleText = stringResource(R.string.hardware__pairing_title))
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
-                .padding(horizontal = 32.dp)
-        ) {
-            BodyM(stringResource(R.string.hardware__pairing_text), color = Colors.White64)
-            FillHeight()
-            // Fixed-width cells so digits replace dots without the row shifting.
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                repeat(PAIRING_CODE_LENGTH) { index ->
-                    val digit = code.getOrNull(index)?.toString()
-                    Box(
-                        contentAlignment = Alignment.Center,
-                        modifier = Modifier.width(PAIRING_CELL_WIDTH)
-                    ) {
-                        Display(
-                            text = digit ?: "•",
-                            color = if (digit != null) Colors.White else Colors.White32,
-                        )
-                    }
-                }
-            }
-            FillHeight()
-        }
-        NumberPad(
-            onPress = { key ->
-                when {
-                    key == KEY_DELETE -> code = code.dropLast(1)
-                    code.length < PAIRING_CODE_LENGTH -> {
-                        code += key
-                        if (code.length == PAIRING_CODE_LENGTH) {
-                            submitted = true
-                            onSubmit(code)
-                        }
-                    }
-                }
-            },
-            includeNavigationBarsPadding = true,
-        )
     }
 }
 
@@ -213,48 +57,4 @@ sealed interface HardwareRoute {
 
     @Serializable
     data object PairingCode : HardwareRoute
-}
-
-private const val PAIRING_CODE_LENGTH = 6
-private val PAIRING_CELL_WIDTH = 32.dp
-
-// Proportions taken from the 375dp-wide Figma frame: 256dp visuals bleeding
-// 84dp off the left edge and 53dp off the right, staggered by 12dp vertically.
-private const val INTRO_IMAGE_SIZE_RATIO = 256f / 375f
-private const val INTRO_TREZOR_BLEED_RATIO = 84f / 375f
-private const val INTRO_LEDGER_BLEED_RATIO = 53f / 375f
-private const val INTRO_IMAGE_STAGGER_RATIO = 12f / 375f
-
-@Preview(showSystemUi = true)
-@Composable
-private fun Preview() {
-    AppThemeSurface {
-        BottomSheetPreview {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .sheetHeight(SheetSize.LARGE, isModal = true)
-                    .gradientBackground()
-            ) {
-                HardwareIntro(onClose = {})
-            }
-        }
-    }
-}
-
-@Preview(showSystemUi = true)
-@Composable
-private fun PairingPreview() {
-    AppThemeSurface {
-        BottomSheetPreview {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .sheetHeight(SheetSize.LARGE, isModal = true)
-                    .gradientBackground()
-            ) {
-                HardwarePairing(onSubmit = {}, onCancel = {})
-            }
-        }
-    }
 }
